Extracted `ModuleBuilder` into its own module
